improved error handling for deleting exercises from a workout

+++ workout_tracker/Core/History/Views/WorkoutDetailsScrollView.swift
@@ -13,14 +13,31 @@ struct WorkoutDetailsScrollView: View {
     let workout: Workout
     
     var body: some View {
-        List {
-            ForEach(viewModel.historyItems) { item in
-                InstanceHistoryRowView(exercise: item.exercise, instance: item.instance)
-            }
-            .onDelete { offset in
-                Task {
-                    guard let idToDelete = viewModel.instances.first(where: { $0.id == viewModel.historyItems[offset.first!].instance.id })?.id else { return }
-                    await viewModel.deleteInstance(by: idToDelete)
+        Group {
+            if !viewModel.historyItems.isEmpty {
+                List {
+                    ForEach(viewModel.historyItems) { item in
+                        InstanceHistoryRowView(exercise: item.exercise, instance: item.instance)
+                    }
+                    .onDelete { offset in
+                        Task {
+                            guard let idToDelete = viewModel.instances.first(where: { $0.id == viewModel.historyItems[offset.first!].instance.id })?.id else { return }
+                            
+                            viewModel.historyItems.remove(atOffsets: offset)
+                            
+                            await viewModel.deleteInstance(by: idToDelete)
+                        }
+                    }
+                }
+            } else {
+                VStack {
+                    Spacer()
+                    
+                    Text("No data for this workout.")
+                        .multilineTextAlignment(.center)
+                        .foregroundColor(Color(.systemGray))
+                    
+                    Spacer()
                 }
             }
         }
